+++ packages/protocol/test/shared/signal/SignalService.t.sol
@@ -107,7 +107,7 @@ contract TestSignalService is CommonTest {
     }
 
     function test_SignalService_proveSignalReceived_revert_malformat_proof() public {
-        // "undecodable proof" is not decodeable into SignalService.HopProof[] memory
+        // "undecodable proof" is not decodable into SignalService.HopProof[] memory
         vm.expectRevert();
         mockSignalService.proveSignalReceived({
             _chainId: 0,