+++ packages/taiko-client/prover/event_handler/interface.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/metadata"
 	ontakeBindings "github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/ontake"
+	pacayaBindings "github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/pacaya"
 	eventIterator "github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/chain_iterator/event_iterator"
 )
 
@@ -24,11 +25,13 @@ type TransitionContestedHandler interface {
 // TransitionProvedHandler is the interface for handling `TaikoL1.TransitionProvedV2` events.
 type TransitionProvedHandler interface {
 	Handle(ctx context.Context, event *ontakeBindings.TaikoL1ClientTransitionProvedV2) error
+	HandlePacaya(ctx context.Context, e *pacayaBindings.TaikoInboxClientBatchesProved) error
 }
 
 // BlockVerifiedHandler is the interface for handling `TaikoL1.BlockVerifiedV2` events.
 type BlockVerifiedHandler interface {
 	Handle(e *ontakeBindings.TaikoL1ClientBlockVerifiedV2)
+	HandlePacaya(ctx context.Context, e *pacayaBindings.TaikoInboxClientBatchesVerified) error
 }
 
 // AssignmentExpiredHandler is the interface for handling the proof assignment expiration.