+++ packages/taiko-client/prover/event_handler/util.go
@@ -7,10 +7,12 @@ import (
 	"math/big"
 	"time"
 
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/log"
 
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/metadata"
+	"github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/pacaya"
 	eventIterator "github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/chain_iterator/event_iterator"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/pkg/rpc"
 )
@@ -113,6 +115,73 @@ func getMetadataFromBlockIDOntake(
 
 	if m == nil {
 		return nil, fmt.Errorf("failed to find BlockProposedV2 event for block %d", id)
+	}
+
+	return m, nil
+}
+
+// getMetadataFromBatchPacaya fetches the batch meta from the onchain event by the given batch id.
+func getMetadataFromBatchPacaya(
+	ctx context.Context,
+	rpc *rpc.Client,
+	batch *pacaya.ITaikoInboxBatch,
+) (m metadata.TaikoProposalMetaData, err error) {
+	callback := func(
+		_ context.Context,
+		meta metadata.TaikoProposalMetaData,
+		_ eventIterator.EndBlockProposedEventIterFunc,
+	) error {
+		if !meta.IsPacaya() {
+			return nil
+		}
+		// Only filter for exact batchID we want.
+		if meta.Pacaya().GetBatchID().Cmp(new(big.Int).SetUint64(batch.BatchId)) != 0 {
+			return nil
+		}
+
+		m = meta
+
+		return nil
+	}
+
+	config, err := rpc.GetProtocolConfigs(&bind.CallOpts{Context: ctx})
+	if err != nil {
+		return nil, fmt.Errorf("failed to get Pacaya protocol configs: %w", err)
+	}
+
+	// Ensure we don't go beyond the current L1 head.
+	l1Head, err := rpc.L1.HeaderByNumber(ctx, nil)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get L1 head: %w", err)
+	}
+	endHeight := new(big.Int).Add(
+		new(big.Int).SetUint64(batch.AnchorBlockId),
+		new(big.Int).SetUint64(config.MaxAnchorHeightOffset()),
+	)
+	if endHeight.Cmp(l1Head.Number) > 0 {
+		endHeight = l1Head.Number
+	}
+
+	iter, err := eventIterator.NewBlockProposedIterator(ctx, &eventIterator.BlockProposedIteratorConfig{
+		Client:               rpc.L1,
+		TaikoL1:              rpc.OntakeClients.TaikoL1,
+		TaikoInbox:           rpc.PacayaClients.TaikoInbox,
+		PacayaForkHeight:     rpc.PacayaClients.ForkHeight,
+		StartHeight:          new(big.Int).SetUint64(batch.AnchorBlockId),
+		EndHeight:            endHeight,
+		OnBlockProposedEvent: callback,
+	})
+	if err != nil {
+		log.Error("Failed to start event iterator", "event", "BatchProposed", "error", err)
+		return nil, err
+	}
+
+	if err := iter.Iter(); err != nil {
+		return nil, err
+	}
+
+	if m == nil {
+		return nil, fmt.Errorf("failed to find BatchProposed event for batch %d", batch.BatchId)
 	}
 
 	return m, nil