+++ packages/taiko-client/prover/proof_producer/compose_proof_producer.go
@@ -38,7 +38,7 @@ type ComposeProofProducer struct {
 	JWT                 string // JWT provided by Raiko
 	PivotProducer       *PivotProofProducer
 	ProofType           ProofType
-	IsOp                bool
+	Dummy               bool
 	DummyProofProducer
 }
 
@@ -74,7 +74,7 @@ func (s *ComposeProofProducer) RequestProof(
 		return err
 	})
 	g.Go(func() error {
-		if s.IsOp {
+		if s.Dummy {
 			proofType = s.ProofType
 			if resp, err := s.DummyProofProducer.RequestProof(opts, batchID, meta, s.Tier(), requestAt); err != nil {
 				return err
@@ -157,7 +157,7 @@ func (s *ComposeProofProducer) Aggregate(
 		return nil
 	})
 	g.Go(func() error {
-		if s.IsOp {
+		if s.Dummy {
 			proofType = s.ProofType
 			resp, _ := s.DummyProofProducer.RequestBatchProofs(items, s.Tier(), s.ProofType)
 			batchProofs = resp.BatchProof
@@ -235,7 +235,11 @@ func (s *ComposeProofProducer) requestBatchProof(
 	}
 
 	if err := output.Validate(); err != nil {
-		return nil, fmt.Errorf("invalid Raiko response (batches: %#v): %w", batches, err)
+		return nil, fmt.Errorf("invalid Raiko response(start: %d, end: %d): %w",
+			batches[0].BatchID,
+			batches[len(batches)-1].BatchID,
+			err,
+		)
 	}
 
 	log.Info(