+++ packages/taiko-client/prover/proof_producer/pivot_proof_producer.go
@@ -15,7 +15,6 @@ import (
 )
 
 // PivotProofProducer generates a pivot proof for the given block.
-// TODO: Can be refactored like ProofProducerPacaya
 type PivotProofProducer struct {
 	Verifier            common.Address
 	RaikoHostEndpoint   string // a prover RPC endpoint
@@ -140,12 +139,17 @@ func (s *PivotProofProducer) requestBatchProof(
 	}
 
 	if err := output.Validate(); err != nil {
-		return nil, fmt.Errorf("invalid Raiko response (batches: %#v): %w", batches, err)
+		return nil, fmt.Errorf("invalid Raiko response(start: %d, end: %d): %w",
+			batches[0].BatchID,
+			batches[len(batches)-1].BatchID,
+			err,
+		)
 	}
 
 	log.Info(
 		"Batch proof generated",
-		"batches", batches,
+		"start", batches[0].BatchID,
+		"end", batches[len(batches)-1].BatchID,
 		"isAggregation", isAggregation,
 		"proofType", proofType,
 		"time", time.Since(requestAt),