+++ packages/taiko-client/prover/proof_producer/zkvm_producer.go
@@ -28,6 +28,8 @@ type ZKvmProofProducer struct {
 	RaikoHostEndpoint   string
 	RaikoRequestTimeout time.Duration
 	JWT                 string // JWT provided by Raiko
+	Dummy               bool
+	DummyProofProducer
 }
 
 // RequestProof implements the ProofProducer interface.
@@ -48,6 +50,10 @@ func (s *ZKvmProofProducer) RequestProof(
 		"coinbase", meta.Ontake().GetCoinbase(),
 		"time", time.Since(requestAt),
 	)
+
+	if s.Dummy {
+		return s.DummyProofProducer.RequestProof(opts, blockID, meta, s.Tier(), requestAt)
+	}
 
 	proof, proofType, err := s.callProverDaemon(ctx, opts, requestAt)
 	if err != nil {
@@ -145,6 +151,9 @@ func (s *ZKvmProofProducer) callProverDaemon(
 	}
 	if output.Data.Status == StatusRegistered {
 		return nil, "", ErrRetry
+	}
+	if output.Data.Status == ErrZkAnyNotDrawn.Error() {
+		return nil, "", ErrZkAnyNotDrawn
 	}
 
 	if !opts.OntakeOptions().Compressed {